Replaced all default 'Float' types with 'Double'.

+++ Sources/GraphPoint/CartesianPoint.swift
@@ -21,7 +21,7 @@ public typealias CartesianPoint = Point
 
 public extension CartesianPoint {
     /// The minimum axis for a `CartesianPlane` that would contain this point.
-    var minimumAxis: Float {
+    var minimumAxis: Double {
         return max(abs(x), abs(y))
     }
 }
@@ -31,8 +31,9 @@ public extension CartesianPoint {
     ///
     /// Uses the mathematical **Law of Sines**.
     ///
-    /// - parameter radius: The straight line distance from the _origin_.
-    /// - parameter degree: The angular degree (0-360), clockwise from the x-axis.
+    /// - parameters:
+    ///   - radius: The straight line distance from the _origin_.
+    ///   - degree: The angular degree (0-360), clockwise from the x-axis.
     /// - returns:A `CartesianPoint` with offsets from the _origin_.
     static func make(for radius: Radius, degree: Degree, clockwise: Bool = true) throws -> CartesianPoint {
         guard degree >= 0.0, degree <= 360.0 else {
@@ -47,10 +48,10 @@ public extension CartesianPoint {
             return .zero
         }
         
-        let rightAngle: Float = 90.0
+        let rightAngle: Double = 90.0
         let sinRight = sin(rightAngle.radians)
-        var rise: Float = 0.0
-        var run: Float = 0.0
+        var rise: Double = 0.0
+        var run: Double = 0.0
         var point: CartesianPoint = .zero
         
         switch clockwise {
@@ -110,9 +111,10 @@ public extension CartesianPoint {
     /// * **c**: calculated based on `degree` and `radius`.
     /// * **a**: supplied via the `point` (x/y based on closest axis)
     ///
-    /// - parameter radius: The straight line distance from the _origin_.
-    /// - parameter degree: The angular degree (0-360), clockwise from the x-axis.
-    /// - parameter modifier: The point used to clip or expand the result. The nearest axis value is used.
+    /// - parameters:
+    ///   - radius: The straight line distance from the _origin_.
+    ///   - degree: The angular degree (0-360), clockwise from the x-axis.
+    ///   - modifier: The point used to clip or expand the result. The nearest axis value is used.
     static func make(for radius: Radius, degree: Degree, modifier: CartesianPoint, clockwise: Bool = true) throws -> CartesianPoint {
         guard degree >= 0.0, degree <= 360.0 else {
             throw GraphPointError.invalidDegree(degree)
@@ -131,28 +133,28 @@ public extension CartesianPoint {
         switch clockwise {
         case true:
             if (degree >= 315) {
-                point.x = sqrtf(powf(radius, 2) - powf(modifier.y, 2))
+                point.x = sqrt(pow(radius, 2) - pow(modifier.y, 2))
                 point.y = modifier.y
             } else if (degree >= 270) {
                 point.x = modifier.x
-                point.y = sqrtf(powf(radius, 2) - powf(modifier.x, 2))
+                point.y = sqrt(pow(radius, 2) - pow(modifier.x, 2))
             } else if (degree >= 225) {
                 point.x = modifier.x
-                point.y = sqrtf(powf(radius, 2) - powf(modifier.x, 2))
+                point.y = sqrt(pow(radius, 2) - pow(modifier.x, 2))
             } else if (degree >= 180) {
-                point.x = -(sqrtf(powf(radius, 2) - powf(modifier.y, 2)))
+                point.x = -(sqrt(pow(radius, 2) - pow(modifier.y, 2)))
                 point.y = modifier.y
             } else if (degree >= 135) {
-                point.x = -(sqrtf(powf(radius, 2) - powf(modifier.y, 2)))
+                point.x = -(sqrt(pow(radius, 2) - pow(modifier.y, 2)))
                 point.y = modifier.y
             } else if (degree >= 90) {
                 point.x = modifier.x
-                point.y = -(sqrtf(powf(radius, 2) - powf(modifier.x, 2)))
+                point.y = -(sqrt(pow(radius, 2) - pow(modifier.x, 2)))
             } else if (degree >= 45) {
                 point.x = modifier.x
-                point.y = -(sqrtf(powf(radius, 2) - powf(modifier.x, 2)))
+                point.y = -(sqrt(pow(radius, 2) - pow(modifier.x, 2)))
             } else if (degree >= 0) {
-                point.x = sqrtf(powf(radius, 2) - powf(modifier.y, 2))
+                point.x = sqrt(pow(radius, 2) - pow(modifier.y, 2))
                 point.y = modifier.y
             }
         case false:
